Change experiments to use one epoch value
Remove the low/high epoch values from the experiment, to make the
experiment granular.

Now that there is only one epoch we no longer need to specify the
epoch parameter separately. It can be passed around as part of
the named tupled object.

To test different values of epochs we need to run separate
experiments. That's how it should be, since epoch is a parameter
for the experiment.