Use latest minor version of each package
Change version specification to use the latest minor version
of each package.

This may cause tests to behave slightly different as the
packages update their versions, but this risk should be
small, since we are asking for minor version updates only.

This change avoids bugs that prevents new installations,
making the project more long-lived, without the nedd for
constant changes to the environment.